compiler error

+++ src/sfsym.rs
@@ -5,7 +5,7 @@ use std::option::Option;
 #[derive(Debug, Clone)]
 pub struct SFSym {
     pub	sym:char,
-	pub count:i32,
+	pub count:usize,
     pub prob:f64,
 	pub coding:String,
 }
